// Mavim-apps/Mavim.frontend/projects/mavim-app/src/shared/field/models/fieldvaluetype.enum.ts
export enum FieldValueType {
	unknown = 'unknown',
	text = 'text',
	multiText = 'multiText',
	number = 'number',
	multiNumber = 'multiNumber',
	decimal = 'decimal',
	multiDecimal = 'multiDecimal',
	boolean = 'boolean',
	date = 'date',
	multiDate = 'multiDate',
	list = 'list',
	relationship = 'relationship',
	multiRelationship = 'multiRelationship',
	relationshipList = 'relationshipList',
	hyperlink = 'hyperlink',
	multiHyperlink = 'multiHyperlink'
}